requirements.txt
smolagents[tools]
ddgs
smolagents[litellm]
uvicorn
fastapi
crewai_tools
selenium